+++ src/js/components/controls/index.js
@@ -1,0 +1,5 @@
+import { Dropdown } from './Dropdown';
+
+export const controls = {
+    Dropdown,
+};
