+++ src/js/components/legacy/index.js
@@ -1,0 +1,6 @@
+import { Dropdown } from './Dropdown';
+
+
+export const legacy = {
+    Dropdown,
+};
